Fix exposure and white balance controls
Correctly handle Exposure mode control:
The value definition for the V4L2 interface for exposure
mode control does not map directly to the UVC control
interface. Add some conversion and configuration fixes.

Enable AWB control:
This is a bit of a hack.
The RaspberryPi V4L2 driver exposes only a menu of pre-defined
color temperatures and two of those are manual and auto.

This control definition does not have any corresponding equivalent
in the UVC specification.

Override the UVC color temperature control interface to map
directly to the V4L2 AWB interface. This will allow controlling
the AWB modes from the host but will only show number values with
no explanation of what the numbers refer to.

0: Manual
1: Auto
2: Incandescent
3: Fluorescent
4: Fluorescent H
5: Horizon
6: Daylight
7: Flash
8: Cloudy
9: Shade

This has been tested only on RaspberryPi 4B HW with the
8 Mpixel v2 camera.

Signed-off-by: Soren Friis <soren.friis@intel.com>